refactor: remove NoteLink component and inline its functionality in UpdateEditor
